requirements.txt
# Genie Backend Dependencies
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP Requests
requests==2.31.0
urllib3==2.0.7

# Environment & Configuration
python-dotenv==1.0.0

# Data Handling
dataclasses-json==0.6.1

# Production Server
gunicorn==21.2.0

# Optional: Database (if you want to upgrade from JSON)
# SQLAlchemy==2.0.23
# psycopg2-binary==2.9.9  # For PostgreSQL
# pymongo==4.6.0  # For MongoDB

# Optional: Monitoring & Logging
# sentry-sdk[flask]==1.38.0
# prometheus-client==0.19.0

# Development (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
